+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 
 - PIC-126 Brush over doesn't trigger update if no collisions occur
 - PIC-120 Property displayOrder of text component should be set to default value instead of 99
+- PIC-129 Bounding rect of a line should never have a height or width of zero
 
 ## [BREAKING] 0.13.0